requirements.txt
pvporcupine
pyaudio
numpy
SpeechRecognition
pyttsx3
pyautogui
requests
plyer
python-dotenv
psutil